Complete task 2

+++ RejseKortet_ext-version/NotEnoughMoneyException.java
@@ -1,5 +1,5 @@
 public class NotEnoughMoneyException extends Exception{
-    public NotEnoughMoneyException(int balance) {
+    public NotEnoughMoneyException(long balance) {
         super("Not enough money in account to check in. Please deposit at least " + Math.abs(balance - 100) + " DKK");
     }
 }
